backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-docx>=1.1.0
python-multipart>=0.0.6